fix table potentially not updating when closing auth email sent modal state
