+++ business-registry-dashboard/app/components/Form/AddBusiness/EmailAuthSent.vue
@@ -1,18 +1,9 @@
 <script setup lang="ts">
-const { t } = useI18n()
 const brdModal = useBrdModals()
-const toast = useToast()
-const affStore = useAffiliationsStore()
 
 defineProps<{
   contactEmail: string
 }>()
-
-async function handleClose () {
-  toast.add({ title: t('form.manageBusiness.toast.emailSent') }) // add success toast
-  await affStore.loadAffiliations() // update table with new affilitations
-  brdModal.close()
-}
 </script>
 <template>
   <div class="flex flex-col gap-4 text-left text-bcGovColor-midGray">
@@ -26,7 +17,7 @@ async function handleClose () {
     <div class="flex justify-center gap-2">
       <UButton
         :label="$t('btn.close')"
-        @click="handleClose"
+        @click="brdModal.close()"
       />
     </div>
   </div>
